LatexGenerator.php: Dump Latex output to debug toolbar
Dump the output to the debug toolbar so there's no need to look for the log file. If we would dump outside of  `if (!$process->isSuccessful())`, the dump would be mixed into the pdf output on success. We could go further by dumping the log file output as well, but we'd need to check if that exists first, and it's pretty much identical to stdout in nonstopmode.